Revert of 9ffadf773d20150300c0697d757c205a913811ba
In a test with ftput (upload) of for example /dev/urandom and/or
/dev/null, after several seconds of seeing successful qcnet_(bg_)startxmit
qcnet_(bg_)complete rounds, it was noticed that the usb_sumbit_urb failed
to deliver its usb_complete_t (qcnet_complete) making the driver
unresponsive (no longer responded on power down, for example).

Lowering the mtu (ifconfig wwan0 mtu 500) seemed to fix the problem.
After reverting the commit the problem was not longer reproducible.

The usb host used was Freescale On-Chip EHCI Host Controller (wfsl-ehci)

 commit 9ffadf773d20150300c0697d757c205a913811ba
 Author: ttuttle <ttuttle@chromium.org>
 Date:   Wed Jul 27 13:11:26 2011 -0400

    CHROMIUM: gobi: Reuse skb buffer in urb instead of copying

    We are needlessly making a copy of each sk_buff we transmit to put into
    the bulk urb.  We can instead reuse the copy in the sk_buff, and hold
    on to our reference until the urb has completed.

    BUG=chromium-os:17703
    TEST=network_3GSmokeTest passed

    Signed-off-by: ttuttle <ttuttle@chromium.org>

    Change-Id: Id9305b46d10bd42f837d85ec423dcdf368e7824e
    Reviewed-on: http://gerrit.chromium.org/gerrit/4826
    Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
    Tested-by: ttuttle <ttuttle@chromium.org>