gobi: Provide netif_wake_queue and netif_stop_queue usage
This adds use of netif_stop_queue and netif_wake_queue to avoid
that the driver just queues everything in startxmit on an end-
less linked list of urbs.

To test this use:
	iperf -s -u -p 5000 -b
	iperf -c <SERVER> -u -p 5000 -b 20M -i 1 -t 1000

The patch also adds setting txqueuelen and mtu using ifconfig
for easier testing.

To test this use:

	ifconfig wlan0 mtu 1500 txqueuelen 10